max value for ticket fix (#128)

+++ app/src/main/java/toluog/campusbash/model/Event.kt
@@ -24,7 +24,7 @@ data class Event(@PrimaryKey var eventId: String = "", var eventName: String = "
                  @Embedded(prefix = "eventVideo_") var eventVideo: Media? = null, var university: String = "",
                  var startTime: Long = 0L, var endTime: Long = 0L, var timeZone: String = "",
                  var placeId: String = "", var tickets: ArrayList<Ticket> = ArrayList(),
-                 @Embedded var creator: Creator = Creator(), var ticketsSold: Int = 0,
+                 @Embedded var creator: Creator = Creator(), var ticketsSold: Long = 0,
                  @Exclude var address: String = ""): Parcelable {
 
     override fun equals(other: Any?): Boolean {
